fpga implementation extracted into fpga.cpp
